Cross-reference to external projects (#123)
This PR enables cross-reference to external projects.
See https://www.sphinx-doc.org/en/master/usage/extensions/intersphinx.html
for full documentation.

Below are 4 different ways to cross-reference documents of an external
project.

```
:doc:`best-practices/software-installation`
:doc:`Any Text <best-practices/software-installation>`
:doc:`seis101:best-practices/software-installation`
:doc:`Any Text <seis101:best-practices/software-installation>`
```

For option 1 and 2, Sphinx will lookup the document in the current
project first, if not found, they lookup in external projects.

For option 3 and 4, Sphinx will force lookup the document in external
projects.

I'm in favor of option 3 and 4. As we're working on multiple projects,
if we have the same file name `best-practices/software-installation.rst`
in multiple projects, Sphinx may link to the wrong one.